refactor: 💡 split types and interfaces and disable rule in esl
