+++ packages/scoreboard/.eslintrc
@@ -8,9 +8,7 @@
   "extends": ["eslint:recommended", "plugin:@typescript-eslint/recommended"],
 
   "rules": {
-    "@typescript-eslint/no-unused-vars": "error",
-    // to enforce using type for object type definitions, can be type or interface
-    "@typescript-eslint/consistent-type-definitions": ["error", "type"],
+    "@typescript-eslint/no-unused-vars": "error"
   },
 
   "env": {
