+++ packages/scoreboard/src/types/match.types.ts
@@ -1,0 +1,10 @@
+export type MatchInitialConfigT = {
+  homeTeam: string;
+  awayTeam: string;
+  homeScore?: number;
+  awayScore?: number;
+};
+export type MatchUpdateScoreT = {
+  homeScore: number;
+  awayScore: number;
+};
